Change how pool is stored in local storage
Save pool to local storage through a function rather than as a side effect to have more control over when it gets set to local storage.
Move the remove blank players function into usePool in order to remove blank players before pool is saved to local storage.
Moved copyPool into usePool to be able to use it in other components and in usePool while avoiding duplication.