Update creating first player inputs
Updated the initial state of the pool slice to have a blank player in the players array. This eliminates the requirement of adding a blank player when the create pool page is rendered, and checking if the players array is empty to add the initial inputs in the playersList component. This simplifies the logic and code required.